+++ projects/Amlogic-ce/packages/tools/u-boot-tools/package.mk
@@ -2,36 +2,39 @@
 # Copyright (C) 2017-present Team LibreELEC (https://libreelec.tv)
 
 PKG_NAME="u-boot-tools"
-PKG_VERSION="2016.03"
-PKG_SHA256="e49337262ecac44dbdeac140f2c6ebd1eba345e0162b0464172e7f05583ed7bb"
-PKG_SITE="https://www.denx.de/wiki/U-Boot"
-PKG_URL="ftp://ftp.denx.de/pub/u-boot/u-boot-$PKG_VERSION.tar.bz2"
-PKG_DEPENDS_HOST="gcc:host"
-PKG_DEPENDS_TARGET="toolchain u-boot-tools:host"
+PKG_VERSION="2025.04"
+PKG_SHA256="439d3bef296effd54130be6a731c5b118be7fddd7fcc663ccbc5fb18294d8718"
+PKG_ARCH="arm aarch64"
 PKG_LICENSE="GPL"
-PKG_LONGDESC="Das U-Boot is a cross-platform bootloader for embedded systems."
+PKG_SITE="https://www.denx.de/wiki/U-Boot"
+PKG_URL="https://ftp.denx.de/pub/u-boot/u-boot-${PKG_VERSION}.tar.bz2"
+PKG_DEPENDS_HOST="gcc:host"
+PKG_DEPENDS_TARGET="toolchain u-boot-tools-aml:host"
+PKG_LONGDESC="Das U-Boot is a cross-platform bootloader for embedded systems (Amlogic tools version)."
 
 make_host() {
   make mrproper
-  make dummy_defconfig
-  make tools-only
+  make tools-only_defconfig
+  make tools-only NO_SDL=1
 }
 
 make_target() {
-  CROSS_COMPILE="$TARGET_PREFIX" LDFLAGS="" ARCH=arm make dummy_defconfig
-  CROSS_COMPILE="$TARGET_PREFIX" LDFLAGS="" ARCH=arm make env
+  make mrproper
+  CROSS_COMPILE="$TARGET_PREFIX" ARCH=arm make tools-only_defconfig
+  CROSS_COMPILE="$TARGET_PREFIX" ARCH=arm make envtools
 }
 
 makeinstall_host() {
   mkdir -p $TOOLCHAIN/bin
-    cp tools/mkimage $TOOLCHAIN/bin
+    cp tools/mkimage $TOOLCHAIN/bin/
+    cp tools/mkenvimage $TOOLCHAIN/bin/
 }
 
 makeinstall_target() {
   mkdir -p $INSTALL/etc
-    cp $PKG_DIR/config/fw_env.config $INSTALL/etc/fw_env.config
+    [ -f $PKG_DIR/config/fw_env.config ] && cp $PKG_DIR/config/fw_env.config $INSTALL/etc/fw_env.config
 
   mkdir -p $INSTALL/usr/sbin
     cp tools/env/fw_printenv $INSTALL/usr/sbin/fw_printenv
-    cp tools/env/fw_printenv $INSTALL/usr/sbin/fw_setenv
+    ln -sf fw_printenv $INSTALL/usr/sbin/fw_setenv
 }
